The table identificator is added

+++ src/SQLQT_UI/Logic/Listeners/MainWindowListener.java
@@ -6,7 +6,11 @@ import java.io.File;
 import java.sql.SQLException;
 import java.sql.Statement;
 
+import javax.swing.JFrame;
+import javax.swing.JOptionPane;
+
 import SQLQT_Utilities.ConnectionToBD;
+import SQLQT_Utilities.DBManager;
 /**
  * 
  * @author Anastasiya Goncharova
@@ -14,7 +18,12 @@ import SQLQT_Utilities.ConnectionToBD;
  */
 public class MainWindowListener implements WindowListener
 {
-
+	JFrame frame;
+	public MainWindowListener(JFrame frame)
+	{
+		this.frame = frame;
+	}
+	
 	@Override
 	public void windowActivated(WindowEvent arg0) {
 		
@@ -24,10 +33,11 @@ public class MainWindowListener implements WindowListener
 	public void windowClosed(WindowEvent arg0) {
 		try {
 			Statement st = ConnectionToBD.con.createStatement();
-			st.executeUpdate("drop table if exists patients");
+			st.executeUpdate("drop table if exists patients" + DBManager.tableId);
 			ConnectionToBD.con.close();
-			File file = new File("patients.db");
+			File file = new File("patients" +  DBManager.tableId + ".db");
 			file.delete();
+			System.exit(0);
 		} catch (SQLException e) {
 			e.printStackTrace();
 		}
@@ -35,14 +45,20 @@ public class MainWindowListener implements WindowListener
 
 	@Override
 	public void windowClosing(WindowEvent arg0) {
-		try {
-			Statement st = ConnectionToBD.con.createStatement();
-			st.executeUpdate("drop table if exists patients");
-			ConnectionToBD.con.close();
-			File file = new File("patients.db");
-			file.delete();
-		} catch (SQLException e) {
-			e.printStackTrace();
+		String ObjButtons[] = {"Yes","No"};
+		int PromptResult = JOptionPane.showOptionDialog(frame,"Are you sure you want to exit?",null,JOptionPane.DEFAULT_OPTION,JOptionPane.WARNING_MESSAGE,null,ObjButtons,ObjButtons[1]);
+		if(PromptResult==JOptionPane.YES_OPTION)
+		{
+			try {
+				Statement st = ConnectionToBD.con.createStatement();
+				st.executeUpdate("drop table if exists patients" + DBManager.tableId);
+				ConnectionToBD.con.close();
+				File file = new File("patients" +  DBManager.tableId + ".db");
+				file.delete();
+				System.exit(0);
+			} catch (SQLException e) {
+				e.printStackTrace();
+			}
 		}
 	}
